@@ -20,16 +20,16 @@ describe('AppComponent', () => {
     expect(app).toBeTruthy();
   });
 
-  it(`should have as title 'ssc'`, () => {
+  it(`should have as title 'ssc-tums'`, () => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.componentInstance;
-    expect(app.title).toEqual('ssc');
+    expect(app.title).toEqual('ssc-tums');
   });
 
   it('should render title', () => {
     const fixture = TestBed.createComponent(AppComponent);
     fixture.detectChanges();
     const compiled = fixture.nativeElement as HTMLElement;
-    expect(compiled.querySelector('.content span')?.textContent).toContain('ssc app is running!');
+    expect(compiled.querySelector('.content span')?.textContent).toContain('ssc-tums app is running!');
   });
 });
